emule: pass full uint64_t to Core::l1_ptr in SWEmuleChip
SWEmuleChip::write_to_device / read_from_device receive uint64_t L1
addresses from the UMD Chip interface. The current implementation casts
to uint32_t before passing to tt_emule::Core::l1_ptr, which:

- Truncates silently for offsets >= 4 GB. Today's DRAM bank sizes
  (~4 GB on Blackhole P150, ~2 GB on Wormhole views) fit, but the
  cast is a latent landmine if larger banks ever appear.
- Drops the safety the wider signature provides (bounds-check abort
  on out-of-range offsets).

Drop the cast and pass through the uint64_t directly. Depends on
tt_emule::Core::l1_ptr having a uint64_t overload (recent change in
the tt-emule repo); narrower uint32_t signature still compiles via
implicit widening but loses the diagnostic value.

Zero behavioral change at current bank sizes.

+++ device/chip/sw_emule_chip.cpp
@@ -90,13 +90,13 @@ tt_emule::Core* SWEmuleChip::get_core(tt_xy_pair core_xy) {
 void SWEmuleChip::write_to_device(CoreCoord core, const void* src, uint64_t l1_dest, size_t size) {
     tt_xy_pair key(core.x, core.y);
     tt_emule::Core* target_core = get_core(key);
-    std::memcpy(target_core->l1_ptr(static_cast<uint32_t>(l1_dest)), src, size);
+    std::memcpy(target_core->l1_ptr(l1_dest), src, size);
 }
 
 void SWEmuleChip::read_from_device(CoreCoord core, void* dest, uint64_t l1_src, size_t size) {
     tt_xy_pair key(core.x, core.y);
     tt_emule::Core* target_core = get_core(key);
-    std::memcpy(dest, target_core->l1_ptr(static_cast<uint32_t>(l1_src)), size);
+    std::memcpy(dest, target_core->l1_ptr(l1_src), size);
 }
 
 // Register I/O forwards to the same memory path — emulated cores have no distinct
